requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
httpx>=0.28.0
websockets>=14.0
openai>=1.60.0
anthropic>=0.43.0
google-genai>=1.5.0
python-multipart>=0.0.18
aiofiles>=24.1.0
